improve peer ID formatting for ed25519 PIDs
Given that we're using the ID multihash for these keys, they don't start with
Qm. As a matter of fact, given the protobuf format, the first few bytes will
always be the same.

To fix this, this patch takes the first two bytes and the last 6 bytes. That
allows us to tell the hash function *and* gives us some unique bytes for
identification.